Crate polkadot_node_core_approval_voting

The Approval Voting Subsystem.

This subsystem is responsible for determining candidates to do approval checks on, performing those approval checks, and tracking the assignments and approvals of others. It uses this information to determine when candidates and blocks have been sufficiently approved to finalize.

Re-exports

pub use criteria::RealAssignmentCriteria;

Modules

approval_db

Approval DB accessors and writers for on-disk persisted approval storage data.

criteria

Assignment criteria VRF generation and checking.

Structs

ApprovalVotingSubsystem

The approval voting subsystem.

AssignmentConfig

Information about the world assignments are being produced in.

Config

Configuration for the approval voting subsystem

Metrics

Approval Voting metrics.

Traits

ApprovalVotingWorkProvider

Trait for providing approval voting subsystem with work.

AssignmentCriteria

A trait for producing and checking assignments.

Functions

start_approval_worker