Module polkadot_node_core_approval_voting::approval_db::v2

Version 2 of the DB schema.

Modules

migration_helpers

Approval DB migration helpers.

Structs

ApprovalEntry

Metadata regarding approval of a particular candidate within the context of some particular block.

BlockEntry

Metadata regarding approval of a particular block, by way of approval of the candidates contained within it.

CandidateEntry

Metadata regarding approval of a particular candidate.

Config

The database config.

DbBackend

DbBackend is a concrete implementation of the higher-level Backend trait

OurAssignment

Details pertaining to our assignment on a block.

StoredBlockRange

A range from earliest..last block number stored within the DB.

Tick

TrancheEntry

Metadata regarding a specific tranche of assignments for a specific candidate.

Enums

Error

Errors while accessing things from the DB.

Functions

load_all_blocks

Return all blocks which have entries in the DB, ascending, by height.

load_block_entry

Load a block entry from the aux store.

load_block_entry_v1

Load a block entry from the aux store in v1 format.

load_blocks_at_height

Load a blocks-at-height entry for a given block number.

load_candidate_entry

Load a candidate entry from the aux store in current version format.

load_candidate_entry_v1

Load a candidate entry from the aux store in v1 format.

load_stored_blocks

Load the stored-blocks key from the state.

Type Definitions

Bitfield

Convenience type definition

Result

Result alias for DB errors.