Module polkadot_node_core_approval_voting::criteria

Assignment criteria VRF generation and checking.

Structs

Config

Information about the world assignments are being produced in.

InvalidAssignment

Assignment invalid.

OurAssignment

Details pertaining to our assignment on a block.

Functions

compute_assignments

Compute the assignments for a given block. Returns a map containing all assignments to cores in the block. If more than one assignment targets the given core, only the earliest assignment is kept.