Module polkadot_node_core_approval_voting::approval_db::v3

Version 3 of the DB schema.

Version 3 modifies the our_approval format of ApprovalEntry and adds a new field pending_signatures for BlockEntry

Re-exports

pub use super::v2::Bitfield;

pub use super::v2::OurAssignment;

pub use super::v2::Tick;

pub use super::v2::TrancheEntry;

Modules

migration_helpers

Approval DB migration helpers.

Structs

ApprovalEntry

Metadata regarding approval of a particular candidate within the context of some particular block.

BlockEntry

Metadata regarding approval of a particular block, by way of approval of the candidates contained within it.

CandidateEntry

Metadata regarding approval of a particular candidate.

CandidateSigningContext

Context needed for creating an approval signature for a given candidate.

OurApproval

Metadata about our approval signature

Functions

load_block_entry_v2

Load a block entry from the aux store in v2 format.

load_candidate_entry_v2

Load a candidate entry from the aux store in v2 format.